Generate resolv.so binding prior to compilation.
In order to provide the documented public interface even though the
binding is generated differently depending on the structure of the
system's resolv.h, it is necessary to use adapter modules. Several of
these are supplied in libresolv-sys/lib.d. These are tried in reverse
alphabetical order until one of them works.  If none works, the reason
for the failure is displayed to make it relatively easy to make a new
one.